+++ src/index.js
@@ -1,12 +1,13 @@
 const {Minimatch} = require('minimatch')
 const sm = require('sitemap')
 const isHTTPS = require('is-https')
-const {unionBy, uniq} = require('lodash')
+const unionBy = require('lodash/unionBy')
+const uniq = require('lodash/uniq')
 const path = require('path')
 const fs = require('fs-extra')
 const AsyncCache = require('async-cache')
-const pify = require('pify')
-const {hostname} = require('os')
+const { promisify } = require('util')
+const { hostname } = require('os')
 
 // Defaults
 const defaults = {
@@ -15,20 +16,25 @@ const defaults = {
   generate: false,
   exclude: [],
   routes: [],
-  cacheTime: 1000 * 60 * 15
+  cacheTime: 1000 * 60 * 15,
+  gzip: false
 }
 
-export default async function sitemap (moduleOptions) {
+module.exports = function module (moduleOptions) {
   const options = Object.assign({}, defaults, this.options.sitemap, moduleOptions)
 
   // sitemap-routes.json is written to dist dir on build mode
   const jsonStaticRoutesPath = path.resolve(this.options.buildDir, path.join('dist', 'sitemap-routes.json'))
 
   // sitemap.xml is written to static dir on generate mode
-  const xmlGeneratePath = path.resolve(this.options.srcDir, path.join('static', options.path))
+  const xmlGeneratePath = path.resolve(this.options.srcDir, path.join(this.options.dir.static || 'static', options.path))
+
+  options.pathGzip = (options.gzip) ? `${options.path}.gz` : options.path
+  const gzipGeneratePath = path.resolve(this.options.srcDir, path.join(this.options.dir.static || 'static', options.pathGzip))
 
   // Ensure no generated file exists
   fs.removeSync(xmlGeneratePath)
+  fs.removeSync(gzipGeneratePath)
 
   let staticRoutes = fs.readJsonSync(jsonStaticRoutesPath, {throws: false})
   let cache = null
@@ -43,7 +49,7 @@ export default async function sitemap (moduleOptions) {
   }
 
   // Extend routes
-  this.extendRoutes(async routes => {
+  this.extendRoutes(routes => {
     // Map to path and filter dynamic routes
     let staticRoutes = routes
       .map(r => r.path)
@@ -69,16 +75,42 @@ export default async function sitemap (moduleOptions) {
 
       // TODO on generate process only and not on build process
       if (options.generate) {
-        // Generate static sitemap.xml
-        const routes = await cache.get('routes')
-        const sitemap = await createSitemap(options, routes)
-        const xml = await sitemap.toXML()
-        await fs.writeFile(xmlGeneratePath, xml)
+        (async () => {
+          // Generate static sitemap.xml
+          const routes = await cache.get('routes')
+          const sitemap = await createSitemap(options, routes)
+          const xml = await sitemap.toXML()
+          await fs.ensureFile(xmlGeneratePath)
+          await fs.writeFile(xmlGeneratePath, xml)
+          if (options.gzip) {
+            const gzip = await sitemap.toGzip()
+            await fs.writeFile(gzipGeneratePath, gzip)
+          }
+        })()
       }
     }
   })
 
-  // Add server middleware
+  if (options.gzip) {
+    // Add server middleware for sitemap.xml.gz
+    this.addServerMiddleware({
+      path: options.pathGzip,
+      handler (req, res, next) {
+        cache.get('routes')
+          .then(routes => createSitemap(options, routes, req))
+          .then(sitemap => sitemap.toGzip())
+          .then(gzip => {
+            res.setHeader('Content-Type', 'application/x-gzip')
+            res.setHeader('Content-Encoding', 'gzip')
+            res.end(gzip)
+          }).catch(err => {
+            next(err)
+          })
+      }
+    })
+  }
+
+  // Add server middleware for sitemap.xml
   this.addServerMiddleware({
     path: options.path,
     handler (req, res, next) {
@@ -110,7 +142,7 @@ function createCache (staticRoutes, options) {
         })
     }
   })
-  cache.get = pify(cache.get)
+  cache.get = promisify(cache.get)
 
   return cache
 }
@@ -136,7 +168,7 @@ function createSitemap (options, routes, req) {
 
   // Create promisified instance and return
   const sitemap = sm.createSitemap(sitemapConfig)
-  sitemap.toXML = pify(sitemap.toXML)
+  sitemap.toXML = promisify(sitemap.toXML)
 
   return sitemap
 }