+++ test/fixture/nuxt.config.js
@@ -1,10 +1,23 @@
+const { resolve } = require('path')
+
 module.exports = {
+  rootDir: resolve(__dirname, '../..'),
   srcDir: __dirname,
   dev: false,
   render: {
     resourceHints: false
   },
-  modules: [
-    ['~/../..', {}]
-  ]
+  modules: ['@@'],
+  sitemap: {
+    exclude: [
+      '/exclude'
+    ],
+    generate: true,
+    gzip: true,
+    hostname: 'http://localhost:3000/',
+    routes: [
+      '1/',
+      'child/1'
+    ]
+  }
 }