+++ test/fixture/pages/_param/index.vue
@@ -1,0 +1,13 @@
+<template>
+<div>
+    /{{ param }}/index
+</div>
+</template>
+
+<script>
+export default {
+  asyncData ({ params }) {
+    return params
+  }
+}
+</script>